Feat: 상품 등록 부분 코드 정리
