+++ src/pages/addProduct/component/infoInputStep/SellerInfoInput.tsx
@@ -8,10 +8,6 @@ const SellerInfoInput = ({ onSellerInfoChange }: SellerInfoInputProps) => {
 	const [sellerComment, setSellerComment] = useState<string>('');
 	const maxCharCount = 500; // 최대 글자 수
 
-	const setSellerInfo = (info: string) => {
-		onSellerInfoChange(info);
-	};
-
 	const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
 		const inputText = e.target.value;
